moved to jflat8

+++ src/main/java/com/tecacet/finance/service/DividendService.java
@@ -1,11 +1,12 @@
 package com.tecacet.finance.service;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
 public interface DividendService {
 
-	Map<LocalDate, Double> getHistoricalDividends(String ticker, LocalDate fromDate, LocalDate toDate)
+	Map<LocalDate, BigDecimal> getHistoricalDividends(String ticker, LocalDate fromDate, LocalDate toDate)
 			throws  StockServiceException;
 
 }
